Подключение к БД через SQLAlchemy и проверки с БД
- Подключение к PostgreSQL через SQLAlchemy (сессия, DBHelper)
- Хелперы для users/movies (exists, get_by_id, count и тд)
- В auth/user/movie тестах добавила проверки согласованности ответа API с БД

+++ models/base_model_movies.py
@@ -14,7 +14,7 @@ class MovieInfoRequest(BaseModel):
     """
     name: str
     imageUrl: str
-    price: float = Field(..., ge=1, description="Цена фильма")
+    price: int = Field(..., ge=1, description="Цена фильма")
     description: str
     location: Location = Field(..., description="Местоположение фильма")
     published: bool
@@ -34,7 +34,7 @@ class MovieInfoResponse(BaseModel):
     description: str
     genreId: int
     imageUrl: Optional[str] = None
-    price: float
+    price: int
     rating: float
     location: Location
     published: bool
